moved download preload script from window to system.
